stm32wrapper4dbg: allows direct R/W in a storage partition
To wrap the fsbl already flashed in SD, let the tool able to open,
read from and write to a block partition.

This permits to directly wrap the fsbl copy in the copy (backup)
boot partition and put the result in the default boot partition:
	stm32wrapper4dbg -s /dev/sdc2 -d /dev/sdc1

Update the example in man page to track this use case.

Signed-off-by: Antonio Borneo <antonio.borneo@st.com>
Change-Id: If0dbf90f49509f7ed82c97d9d84d644918ef3a56
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/stm32wrapper4dbg/+/161661
Reviewed-by: CITOOLS <smet-aci-reviews@lists.codex.cro.st.com>