stm32wrapper4dbg: fix handling of Cortex-M fsbl
The bootrom ignores the header's entry point and uses the reset
vector at fixed address at the beginning of the fsbl to start the
fsbl after a CPU reset.

Verify that the fsbl has the same address for header's entry point
and for reset vector's reset address.
Modify the reset vector to let it point to the wrapper.

Signed-off-by: Antonio Borneo <antonio.borneo@foss.st.com>
Change-Id: I35e91dc3935338cfd9ac071a32eb8de6cb610f00
Reviewed-on: https://gerrit.st.com/c/mpu/oe/st/stm32wrapper4dbg/+/368516